移除 Docker 部署与陈旧文档,新增需求/差距分析文档;整合安全认证与前后端优化

+++ src/main/java/com/project/platform/dto/CurrentUserDTO.java
@@ -3,6 +3,8 @@ package com.project.platform.dto;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CurrentUserDTO {
     @NotNull(message = "用户 id 不能为空")
@@ -14,7 +16,7 @@ public class CurrentUserDTO {
     private String tel;
     private String email;
     /**
-     * 余额
+     * 余额(仅作快照展示,实际余额一律实时读 DB)
      */
-    private Float balance;
+    private BigDecimal balance;
 }
